Move `imaging_element` to `imaging_rec`
